pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cuvarbase"
dynamic = ["version"]
description = "Period-finding and variability on the GPU"
readme = "README.rst"
requires-python = ">=3.9"
license = "GPL-3.0-or-later"
license-files = ["LICENSE.txt"]
authors = [
    {name = "John Hoffman", email = "johnh2o2@gmail.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: C",
    "Programming Language :: C++",
    "Topic :: Scientific/Engineering :: Astronomy",
]
dependencies = [
    "numpy>=1.6",
    "scipy",
    "pycuda>=2017.1.1,!=2024.1.2",
    "nvmath-python",
]

[project.optional-dependencies]
test = ["pytest", "nfft", "matplotlib", "astropy"]

[project.urls]
Homepage = "https://github.com/johnh2o2/cuvarbase"
Documentation = "https://johnh2o2.github.io/cuvarbase/"
Source = "https://github.com/johnh2o2/cuvarbase"
Issues = "https://github.com/johnh2o2/cuvarbase/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cuvarbase = ["kernels/*.cu"]

[tool.setuptools.dynamic]
version = {attr = "cuvarbase.__version__"}

[tool.pytest.ini_options]
testpaths = ["tests"]
